pcb, add small board to change USB connectors 90 degrees
The Pi Zero has two microUSB OTG connectors, and the one for data is
inconveniently located in the middle of the board. There's no adapter
that I can easily find that will take this around 90 degrees in the
dimensions of the larger hat board (77x77mm) - so create a small PCB
that connects to both microUSB ports and provides a type-A female socket
facing the right direction - with enough space for a Noolec full-size
SDR.

The microUSB plug into the power socket on the RPi is not connected
(except to ground) - it's just there for mechanical stability. I hoped
to provide a breakout but it's really tight. Maybe in a future revision.

There's an LED on there to provide a power indicator. I'm not sure if
I'll end up populating it when assembled, but it might be useful for
debugging.

Designed for the microUSB plug available at elecbee as EB-500-4019
(literally the only reasonable one I could find) and the type-A socket
from Valcon (CSP-USC16-TR). Footprints for both are included in the
commit, along with a footprint for a much better USB-C connector for the
hat board (the Molex originally specified is extremely difficult to
solder).