aircraft, migrate from local server to subprocess
Aircraft tracking is currently done via a dump1090-mutability daemon
running on the local machine, which we connect to over a TCP socket.
I want to be able to switch frequencies in software, so this change
migrates to using dump1090-fa (the flightaware fork at
https://github.com/flightaware/dump1090) as a subprocess.

Rather than relying on dump1090 to parse the messages for us, it dumps
them in a raw form and we use pyModeS
(https://github.com/junzis/pyModeS) to parse them. This should give us
more flexibility to parse different message types in future.

Tested (in SF) on 978 and 1090Mhz, seems to work.