deploy, add an ansible playbook to deploy paradar
Add an ansible playbook to deploy paradar to a stock Raspbian image. It
ensures a ton of default packages are purged, installs dependencies, and
sets up a virtualenv under /opt/paradar.

Systemd is configured to ensure paradar starts on boot, and is restarted
if necessary.

There's probably more things that can be cleaned up here - and things
like wifi config need work. But this is a reasonable first pass.