gps, rip out serial interface and replacewith gpsd
Oh god, that GPS serial interface code. So janky, so unreliable. It
didn't help that I didn't realise getty was running on the same serial
line and just randomly injecting noise.

It turns out gpsd is a thing: a lightweight daemon that holds the
last-seen position from the GPS. It's also useful for Chrony, so the
system has a reasonably accurate clock (the serial interface means it's
not super-accurate, but it's within a second).

Rip out all that nasty serial code and replace it with a tiny call to
the gpsd python library.