paradar, improve imports and add Config class
Add a Config class to allow lookups against the state of the
user-configurable DIP switch. The mapping from GPIO (switch #) to
function is held in one place, and the functions are expressed as
static methods on the class.

To facilitate testing of this kind of code, fix the module imports so
they check whether we're actually running on a Pi. If not, don't import
things that depend on that hardware. This won't let the program run but
permits unit testing.