case, improve lid geometry
The lid was always a bit of a hack - I did the countersunk cutouts for
the screws after the shell, which was a mistake. Rework the lid to
simplify the geometry, which should hopefully make it more reliable to
print.

Also - shift the power switch posts even further south, and enlarge the
screw posts to give more material to fit screws into. Remove the
bevelling since it was adding a ton of complexity, but still bevel
top/bottom exterior vertices so the device is nice to hold.