layout usb v1.1, adjust positioning and add alternative layout
The v1.0 layout didn't fit, sadly. The current male connector positions
are a little too high on the board.  Bring them down by 2-3mm, and add
vias to allow connection of hookup wire if the male connectors don't
work out.

Remove the power LED and resistor, it's not useful and we don't have a
huge amount of space.

Add an alternative version of the board that can live on the same panel
- it has the female USB connector on the bottom to allow for a different
stack of components in the case (PI zero much higher). This is a
fallback if the current stack doesn't work.

Remove the solder mask from the mounting vias.

As-built gerbers included for both board versions.