paradar gdl90, add encoder for traffic messages
A lot of the fields are stubbed out for now, but this is a first pass at
generating a traffic report in GDL90 from a received aircraft position.

The ADS-B receiver code will need to be improved before we get things like
track and altitude, but this is a place to start.

Not yet tested substantially.