paradar, adjust nice for dump1090 and main process
Renice the main process to 5, and the subprocess to 10. This should allow
system things in the background to keep up, and deprioritises traffic
reception. ADS-B sends updates really regularly (every second) so it's not
worth spending huge cycles receiving every packet. Better to rapidly display
the information we have to the user.