case, add gcode files with appropriate settings to print the cases
Through a long period of iteration (I'm new to 3D printing) I've finally
found a series of settings that print the case and lid with good quality
in reasonable time.

Include rendered gcode files for my printer, an AnyCubic Mega S. The
-base and -lid files are the appropriately sized base and lid
respectively. These fit the board and components well, but don't leave
room for screws to fit through the case :S

The -scaled versions naively add 2mm to the X and Y dimensions. They
work well, could be snugger, but are good enough!

If I have time, I'll redo the blender model to scale the shell and
reposition the internal components in the new volume. For now, this
works ok.