layout v1.5, update pcb
Add the two new sensors to the PCB on the SPI bus, and remove
unnecessary headers to save space. Remove the FLARM header.

Relocate capacitors and the 3v3 regulator that collide with the SDR
during assembly - tuck them just under the rpi, they should be
low-profile enough.

Change the battery connector to match the connectors supplied with the
batteries, so I don't have to change them.

Add soldermask over the mounting holes where the PCB hits the SDR (the
lower half). Tidy up the soldermask a little bit and increase copper
clearance in tight areas.

Remove copper through-hole plating on the USB connector mounting holes
to improve manufacturing reliability.

Add a BOM spreadsheet negotiated with PCBWay - I'll supply the yellow
LEDs and USB-C connectors (these should be subbed out in a future
version).

Trim the corners of the board (where they come up against the case) to
provide more space for the lid screws to fit.