layout v1.5, relocate thermistor away from power supply
The battery charger gets incredibly hot when the battery is charging and
the system is under load - I guess we're pushing it's max current limit
@ 2-3 amps.

To avoid it shutting of unnessarily, relocate the thermistor south so
it's measuring the battery temperature, not the board temperature.

The charge controller has a die thermal cutoff @ ~125 degrees, so it
should handle its own thermal management ok. The thermistor is just
there to protect the battery.