paradar network, stop wifi restarting
The wifi service monitors the GPIO pins for the DIP switches, and relies on the
RPi's internal pullups to stop the switches cycling. This is fine until the
main paradar service stops or crashes - in which case the signal handler calls
GPIO.cleanup() and destroys the pullup configuration.

Reconfigure pullups at the start of every network service loop, to make sure
that when we read the GPIOs we have the correct values.