paradar, print traceback for all threads on signal
To help debug lockups, print a traceback on all threads when we receive
SIGUSR1. I found a service on one unit hung waiting on a recv, and it was
impossible to figure out why. Hopefully this will help if I can reproduce.

Reorder the startup code a little, and track threads in a data structure to
make it easier to print them later.