pcb, slight tweaks to right angle USB
The type-b right angle USB board is the one we want.. but because I bet
on the type a I don't have enough. So I need to order yet more board ... :'(

It's an excuse to try panelising for the first time. This commit
includes a tweaked rightangle board:
 - narrower shaft with more room for the SDR
 - USB connectors a little closer to the edge so they fit better
 - reroute some traces accordingly

Changes are very minor so I didn't bump the version on the board.

Add as-built individual boards, plus a panelised version.