case, create STL files for new rapid assembly
After some experimentation, I've decided to switch to the right-angle
v1.1B USB PCB - ie. the board is reversed and up high. This means the
board can clip neatly to the raspberry pi using a couple of USB-male
connectors. This vastly speeds up assembly, but it needs some changes to
the case.

The USB board is slightly more separated from the raspberry pi than
planned - 6mm instead of 3mm. This means the case needs to be larger in
the Y dimension and slightly shorter in the X dimension.

The board is also up much higher, so the lower cutaway from the
bottom-right screw post doesn't help us. I experimented (009-*) with
removing the bottom-right screw post, to help the SDR fit - the
right-angle PCB collides with it otherwise. Unfortunately this isn't
great - the lid pops up and looks jank.

So in 010-* I expanded the case slightly in the X dimension and trimmmed
the screw posts back a little. This will fit with some modifications to
the USB board (with a file if I don't get any new ones made up).

All the bevelling is gone because that was a dumb idea.

I'm also printing in PETG now. This is way better in many respects (less
temperature sensitive) but it's real hard to get a clean print. Include
some gcodes which get reasonable prints on my Anycubic Mega-S, but
there's more work to do here to get prints that look good.