first pass at a sensor interface in go
The current python compass library is pretty slow, and uses bitbanged SPI. This
change fixes the hardware SPI (via reordering of config.txt, which apparently
matters) and rewrites the hardware drivers in golang to see if it reduces resource usage.

The concept is that a golang binary manages communications with the compass,
pressure sensor and acceleromter, and returns a tilt-compensated azimuth plus
altitude (and other motion data as required) back to the python library.

Who knows, maybe golang is so much easier for this I rewrite the whole thing.
Certainly it's going faster than python did originally.

Current status: sensor.go communications with the compass and returns raw
readings for x/y/z.