go, read and parse pressure sensor data
Read the pressure sensor and calculate the altitude from the temperature, and
an assumed mean sealevel pressure. Note that the temperature reading is
particularly inaccurate - the boards gets warm in the case, and it can read
20-25 degrees high. An offset is added to try and correct with this so the
reading is somewhat accurate.

Move compass.go into it's own package to avoid namespace collisions.