system, update cmdline and config.txt for new kernel
The new kernel hangs on boot with the frequency options. Set the root partition
using PARTUUID to be sure (unsure if that was the root cause also).

While I'm here, drive all three CS lines high during boot to avoid
accelerometer lockups (my guess is that it's getting garbage during the boot
process).