ahrs, improve behaviour when device inverted
When the device is flipped over, the azimuth sent to the display needs to be
-ve and rotated 180°. There's also issues during the process of flipping - when
it's standing on its edge, the compass readings are crazy because the levelling
process goes asymptotic. Freeze the display and discard measurements during these periods.

All in all, this is producing a really nice display.