sensor, clean up branch and add compiled binary
Most notably, fix a bug with float parsing in the altimeter.

CHecking in a binary is a bit jank, but is a stopgap until there's a real build
system.

Note this has not been tested on boards without an IMU (ie. v1.4); it will
likely need a better moving average or similar to reduce noise.