compass, tweak calibration routine
Test and tweak the calibration routine, it seems to work ok. It's not doing as
much as I'd like to fix the hard iron interference, and maybe a better
ellipsoid algorithm would help here. Still, the compass is accurate within
~20° against a good quality magnetic compass.

Rename the compass calibration protobuf to avoid a warning at startup
(apparently names can't be duplicate in the same binary).

Check in the calibration binary for the same reasons as main.