sensor, set beta in AHRS back to 0.1
I set this parameter based on Madwick's paper, which says it should be equal to
the average of the gyro calibration readings multiplied by a constant. However,
this causes serious drift in roll and pitch when the device is left on an
uneven surface for long periods. Setting beta to the commonly-found value of
0.1 seems to resolve this.